+++ python/demo.py
@@ -6,7 +6,7 @@ from rapidocr import RapidOCR
 engine = RapidOCR()
 
 img_url = "https://img1.baidu.com/it/u=3619974146,1266987475&fm=253&fmt=auto&app=138&f=JPEG?w=500&h=516"
-result = engine(img_url, return_word_box=True, return_single_char_box=True)
+result = engine(img_url)
 print(result)
 
 result.vis("vis_result.jpg")
